Relocate useInitialisationStatus to common
