+++ client/packages/common/src/api/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './useQuerySelector';
 export * from './useFieldsSelector';
+export * from './useInitialisationStatus';
