+++ client/packages/common/src/api/operations.graphql
@@ -1,0 +1,6 @@
+query initialisationStatus {
+  initialisationStatus {
+    status
+    siteName
+  }
+}
