+++ client/packages/host/src/api/hooks/utils/index.ts
@@ -1,11 +1,9 @@
 import { useHostApi } from './useHostApi';
-import { useInitialisationStatus } from './useInitialisationStatus';
 import { useSyncStatus } from './useSyncStatus';
 import { useSyncInfo } from './useSyncInfo';
 
 export const Utils = {
   useHostApi,
-  useInitialisationStatus,
   useSyncStatus,
   useSyncInfo,
 };
