+++ client/packages/host/src/api/operations.graphql
@@ -11,10 +11,6 @@ query syncSettings {
   }
 }
 
-query initialisationStatus {
-  initialisationStatus
-}
-
 query displaySettings($input: DisplaySettingsHash!) {
   displaySettings(input: $input) {
     customTheme {
